fbs/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-redis==5.4.0
psycopg2-binary==2.9.9
requests==2.31.0
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
python-decouple==3.8
whitenoise==6.6.0
Pillow==10.1.0
